Исправлены замечания 1) скорректировано поведение прогресс бара 2) удален usecase содержащий тип Intent и все что с ним было связано, код с интентами вернул в активити 3) увеличил размер картинки альбома в форме плейера
