--- app/src/main/java/com/practicum/playlistmaker/domain/usecase/GetIntentsUsaCase.kt
@@ -1,12 +1,0 @@
-package com.practicum.playlistmaker.domain.usecase
-
-import com.practicum.playlistmaker.domain.storage.interfaces.IntentUser
-
-
-class GetIntentsUsaCase(private val intentUser: IntentUser) {
-
-    fun getUserIntent(): IntentUser {
-        return intentUser
-    }
-
-}
