+++ app/src/main/res/layout/activity_player.xml
@@ -54,8 +54,8 @@
 
         <ImageView
             android:id="@+id/iv_playerBanner"
-            android:layout_width="280dp"
-            android:layout_height="280dp"
+            android:layout_width="312dp"
+            android:layout_height="312dp"
             android:layout_marginTop="15dp"
             android:scaleType="centerCrop"
             android:src="@drawable/empty_image"
